ai-dev-system/requirements.txt
mcp==1.26.0
anyio>=4.13.0,<5.0.0
keyboard>=0.13.5,<1.0.0
opencv-python>=4.13.0,<5.0.0
pillow>=12.1.0,<13.0.0
psutil>=7.2.0,<8.0.0
pyautogui>=0.9.54,<1.0.0
PyYAML>=6.0.2,<7.0.0
pytest>=8.4.0,<9.0.0
pywinauto>=0.6.9,<0.7.0
